Cleanup + Logger
This is a general cleanup, including code, comments and the addition of
the logger.

The main is much cleaner now, due to the addition of  the GetMorpher
class.

This is probably where I should have started with the initial commit :)

TODO: Next step is make the hierarchy a bit stronger, this will be done
on top of the NoteMorpherRule
(to deal with start and end)